pyproject.toml
[project]
name = "book-club"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "markdown>=3.8.2",
    "social-auth-app-django>=5.5.1",
]
